pyproject.toml
[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[project]
name = "moist-bot"
version = "0.1.0"
description = "moist"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.14",
    "aiosqlite>=0.17.0",
    "asqlite>=2.0.0",
    "asyncpraw>=7.8.1",
    "charset-normalizer>=3.4.2",
    "colorama>=0.4.6",
    "discord-ext-menus>=1.0.0a31",
    "discord-py>=2.6.0a5228",
    "jishaku>=2.6.0",
    "mypy-extensions>=1.1.0",
    "nekos-py>=1.1.0",
    "numpy>=2.3.2",
    "orjson>=3.11.1",
    "parsedatetime>=2.6",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "pyperclip>=1.9.0",
    "python-dateutil>=2.9.0.post0",
    "pytube>=15.0.0",
    "requests>=2.32.4",
    "typing-extensions>=4.14.1",
]

[tool.uv.sources]
discord-py = { git = "https://github.com/Rapptz/discord.py", rev = "master" }
discord-ext-menus = { git = "https://github.com/Rapptz/discord-ext-menus", rev = "fbb8803779373357e274e1540b368365fd9d8074" }

[project.scripts]
app = "moist_bot.__main__:main"
wf-market-checker = "moist_bot.__main__:main"

[tool.pyright]
typeCheckingMode = "strict"

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
preview = true
select = ["ALL"]
ignore = [
    "E501",
    "T201",
    "EXE002",
    "ANN204",
    "RUF006",
    "RUF029",
    "S404",
    "S602",
    "TC001",  # False positives with pydantic
    "TC002",  # False positives with pydantic
    "TC003",  # False positives with pydantic
    "TD003",
    "FIX002",
    "W293",
    "E303",
    "G004",
    "ERA",

    # discord.py moment
    "ANN201",
    "ANN202",
    "RET503",
    "EM101",
    "FBT001",
    "FBT002",
    "C901",
    "PLR0914",
    "PLR0915",

    # Umbra preset
    "ANN401",
    "ASYNC109",
    "COM812",
    "CPY001",
    "D",        # not needed here
    "DOC",
    "INP001",   # due to how I load modules this results in annoyances
    "PD011",    # this is not a numpy codebase
    "PERF203",
    "PLC0414",  # pyright ruling for `as` imports needed
    "PLR0912",
    "PLR0913",
    "PLR2004",  # I hate this rule
    "PLR6301",
    "Q000",
    "RUF001",
    "RUF009",
    "SIM105",
    "TRY003",   # over-eager rule
    "TRY301",   # unrealistic rule
    "UP034",
]
unfixable = [
    "E501", # line length handled in other ways by ruff format
    "ERA",  # Don"t delete commented out code
]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
split-on-trailing-comma = true
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false
parametrize-names-type = "csv"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
